update wechat model path

+++ QRUtil/WndLogic.cpp
@@ -797,9 +797,9 @@ void MainWnd::SetHintText(Gdiplus::Graphics& graphics, std::string text)
 #endif
 
 	std::wstring wide;
-	auto enc = Encoding::getEncoding((const uchar*)text.c_str(), text.size());
+	auto enc = Encoding::getEncoding((const uchar*)text.c_str(), (uint)text.size());
 	uchar dst[1000];
-	auto len = Encoding::decode((short*)dst, 1000, (const uchar*)text.c_str(), text.size(), enc);
+	auto len = Encoding::decode((short*)dst, 1000, (const uchar*)text.c_str(), (uint)text.size(), enc);
 	wide = wstring((wchar_t*)dst, len);
 	// wide = converter.from_bytes(text.c_str(), text.c_str()+text.size());
